declare type explicitly

+++ instruction.go
@@ -894,7 +894,7 @@ var Instructions = map[InstructionCode]func(cpu *CPU, raw, pc uint64) Exception{
 	SRLIW: func(cpu *CPU, raw, _ uint64) Exception {
 		i := ParseI(raw)
 		shamt := i.Imm & 0b1_1111
-		cpu.XRegs.Write(i.Rd, uint64(int64(int32(cpu.XRegs.Read(i.Rs1)>>shamt))))
+		cpu.XRegs.Write(i.Rd, uint64(int64(int32(uint32(cpu.XRegs.Read(i.Rs1))>>shamt))))
 		return ExcpNone
 	},
 	SRAW: func(cpu *CPU, raw, _ uint64) Exception {
